Add the ability to choose a firewall resolution method
A user is now able to choose how to resolve the firewall:

Manual (default): show a WebView dialog, the user resolves the firewall by himself, and the dialog closes when the firewall is resolved.

Auto (old behavior): a WebView in the background will try to resolve the firewall.

Auto then manual: if auto resolution fails, the manual one is invoked.