Remove the title check from the onPageFinished method due to possible title's localization and rewrite the rest of the code to comply with this change
These changes are needed because the old resolution method relies on checking if the page's title is one of the titles that Cloudflare can use. But Cloudflare can use localized titles, and this breaks resolution logic. The code checks the page's title, sees that the title is not one of Cloudflare's titles because it is localized, and decides that resolution has finished. Then the code tries to collect the cookie, but the cookie is not available because the resolution has not actually finished. That leads to resolution failure.